Fix compiler warning

+++ phrase-translator-cli/phrasepattern.cpp
@@ -1,9 +1,9 @@
 #include "phrasepattern.h"
 
-PhrasePattern::PhrasePattern(POSTag *tags, int tags_size)
+PhrasePattern::PhrasePattern(POSTag *tags, size_t tags_size)
 {
     m_vec.reserve(tags_size);
-    for (int i = 0; i < tags_size; i++) {
+    for (size_t i = 0; i < tags_size; i++) {
         m_vec.push_back(tags[i]);
     }
 }
